Fixed: Actualizar configuracion antes de recargar pagina
